Fetch IDP Admin1 Data

Retrieve IDP data at Admin 1 level based on specified parameters. At least one of the following parameters must be provided: Operation, CountryName, or Admin0Pcode.

Usage

get_idp_admin1_data(
  Operation = NULL,
  CountryName = NULL,
  Admin0Pcode = NULL,
  Admin1Name = NULL,
  Admin1Pcode = NULL,
  FromReportingDate = NULL,
  ToReportingDate = NULL,
  FromRoundNumber = 0,
  ToRoundNumber = 0
)

Arguments

Operation

Optional; Name of the DTM operation for which the data was collected.

CountryName

Optional; Name of the country where the data was collected.

Admin0Pcode

Optional; Country code (ISO 3166-1 alpha-3).

Admin1Name

Optional; Name of level 1 administrative boundaries.

Admin1Pcode

Optional; Place code of level 1 administrative boundaries.

FromReportingDate

Optional; Start date for the reporting period (format: 'YYYY-MM-DD').

ToReportingDate

Optional; End date for the reporting period (format: 'YYYY-MM-DD').

FromRoundNumber

Optional; Starting round number for the data collection range.

ToRoundNumber

Optional; Ending round number for the data collection range.

Value

A data frame containing the IDP Admin1 data matching the specified criteria.

Examples

# Fetch IDP data at Admin Level 1
idp_admin1_df <- get_idp_admin1_data(CountryName='Sudan', Admin1Name="Blue Nile")
head(idp_admin1_df)
#>     id              operation admin0Name admin0Pcode admin1Name admin1Pcode
#> 1 1663        Darfur conflict      Sudan         SDN  Blue Nile        SD08
#> 2 1664        Darfur conflict      Sudan         SDN  Blue Nile        SD08
#> 3 1665        Darfur conflict      Sudan         SDN  Blue Nile        SD08
#> 4 1666        Darfur conflict      Sudan         SDN  Blue Nile        SD08
#> 5 6290 Armed Clashes in Sudan      Sudan         SDN  Blue Nile        SD08
#> 6 6291 Armed Clashes in Sudan      Sudan         SDN  Blue Nile        SD08
#>   numPresentIdpInd       reportingDate yearReportingDate monthReportingDate
#> 1            81693 2021-03-30T00:00:00              2021                  3
#> 2           130958 2021-12-31T00:00:00              2021                 12
#> 3           151156 2022-01-31T00:00:00              2022                  1
#> 4           152656 2022-03-30T00:00:00              2022                  3
#> 5              260 2023-04-28T00:00:00              2023                  4
#> 6              715 2023-05-07T00:00:00              2023                  5
#>   roundNumber assessmentType
#> 1           2             BA
#> 2           3             BA
#> 3           4             BA
#> 4           5             BA
#> 5           2             BA
#> 6           3             BA